requirements.txt
aiogram==3.12.0
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
# Библиотеки для генерации изображений
# Удалено: diffusers
# Удалено: torch==2.2.2
# Удалено: torchvision==0.17.2
transformers==4.44.2
accelerate==0.34.1
safetensors==0.4.3
Pillow==10.4.0

# Новые зависимости для безопасности и производительности
pydantic==2.5.0
redis==5.0.1
structlog==23.2.0
tenacity==8.2.3
prometheus-client==0.19.0

# Зависимости для платежей
cryptography==41.0.7
pyjwt==2.8.0